Move LazyList to new verta._bases
