+++ client/verta/verta/registry/entities/_models.py
@@ -4,13 +4,13 @@ from __future__ import print_function
 
 import copy
 
-from ..._protos.public.registry import RegistryService_pb2 as _RegistryService
-from ..._internal_utils import _utils
+from verta._protos.public.registry import RegistryService_pb2 as _RegistryService
+from verta._bases import _LazyList
 
 from ._model import RegisteredModel
 
 
-class RegisteredModels(_utils.LazyList):
+class RegisteredModels(_LazyList):
     """Collection object for finding registered models.
 
     There should not be a need to instantiate this class directly; please use
